move ArticleForm to formcontext and optimize imports
