+++ portality/view/forms.py
@@ -4,68 +4,14 @@ All the various simple forms that we need in the system
 FIXME: probably we need to think more about how we organise forms - maybe a module with different files for different
 features, and a bit of refactoring on the organisation of the formcontext stuff to be consistent
 '''
-from datetime import datetime
 import re
-from flask_login import current_user
-
 
 from wtforms import Form, validators
-from wtforms import StringField, TextAreaField, FormField, FieldList, HiddenField
+from wtforms import StringField, TextAreaField, HiddenField
 
-from portality.core import app
 from portality import models
-from portality.formcontext.validate import ThisOrThat, OptionalIf, MaxLen
-from portality.formcontext.fields import DOAJSelectField
-from portality import regex
-from portality.formcontext.choices import Choices
+from portality.formcontext.validate import OptionalIf, MaxLen
 
-##########################################################################
-## Forms and related features for Article metadata
-##########################################################################
-
-DOI_REGEX = regex.DOI_COMPILED
-DOI_ERROR = 'Invalid DOI. A DOI can optionally start with a prefix (such as "doi:"), followed by "10." and the remainder of the identifier'
-ORCID_REGEX = regex.ORCID_COMPILED
-ORCID_ERROR = "Invalid ORCID iD. Please enter your ORCID iD as a full URL of the form https://orcid.org/0000-0000-0000-0000"
-
-# use the year choices in app.cfg or default to 15 years previous.
-start_year = app.config.get("METADATA_START_YEAR", datetime.now().year - 15)
-
-YEAR_CHOICES = [(str(y), str(y)) for y in range(datetime.now().year + 1, start_year - 1, -1)]
-MONTH_CHOICES = [("1", "01"), ("2", "02"), ("3", "03"), ("4", "04"), ("5", "05"), ("6", "06"), ("7", "07"), ("8", "08"), ("9", "09"), ("10", "10"), ("11", "11"), ("12", "12")]
-INITIAL_AUTHOR_FIELDS = 3
-
-class AuthorForm(Form):
-    name = StringField("Name", [validators.Optional()])
-    affiliation = StringField("Affiliation", [validators.Optional()])
-    orcid_id = StringField("ORCID iD", [validators.Optional(), validators.Regexp(regex=ORCID_REGEX, message=ORCID_ERROR)])
-
-class ArticleForm(Form):
-    title = StringField("Article Title", [validators.DataRequired()])
-    doi = StringField("DOI", [OptionalIf("fulltext"), validators.Regexp(regex=DOI_REGEX, message=DOI_ERROR)], description="(You must provide a DOI and/or a Full-Text URL)")
-    authors = FieldList(FormField(AuthorForm), min_entries=1) # We have to do the validation for this at a higher level
-    abstract = TextAreaField("Abstract", [validators.Optional()])
-    keywords = StringField("Keywords", [validators.Optional()], description="Use a , to separate keywords") # enhanced with select2
-    fulltext = StringField("Full-Text URL", [OptionalIf("doi"), validators.URL()], description="(The URL for each article must be unique.  You must provide a Full-Text URL and/or a DOI)")
-    publication_year = DOAJSelectField("Year", [validators.Optional()], choices=YEAR_CHOICES, default=str(datetime.now().year))
-    publication_month = DOAJSelectField("Month", [validators.Optional()], choices=MONTH_CHOICES, default=str(datetime.now().month) )
-    pissn = DOAJSelectField("Journal ISSN (print version)", [ThisOrThat("eissn")], choices=[]) # choices set at construction
-    eissn = DOAJSelectField("Journal ISSN (online version)", [ThisOrThat("pissn")], choices=[]) # choices set at construction
- 
-    volume = StringField("Volume Number", [validators.Optional()])
-    number = StringField("Issue Number", [validators.Optional()])
-    start = StringField("Start Page", [validators.Optional()])
-    end = StringField("End Page", [validators.Optional()])
-
-    def __init__(self, *args, **kwargs):
-        super(ArticleForm, self).__init__(*args, **kwargs)
-        try:
-            self.pissn.choices = Choices.choices_for_article_issns(current_user)
-            self.eissn.choices = Choices.choices_for_article_issns(current_user)
-        except:
-            # not logged in, and current_user is broken
-            # probably you are loading the class from the command line
-            pass
 
 ##########################################################################
 ## Editor Group Forms
